Fix undo glitch
This should take care of the glitchy behavior of the undo

There was an assumption being made where if C was held down at all then
we should start undoing moves.  Unfortunately, the else-if branch for C
being held down was low priority.  Basically the two ifs didn't agree
with each other